Add PERSIST_COOKIES module-level constant.
Cookie persistence is no longer the default behaviour,
instead one has to instead pass in a CookieJar instance
via the kwarg, or set the module-level constant
PERSIST_COOKIES to True and the HTTPClient will use the
custom PersistFileCookieJar instead.

This ensures the wrapper is kept by default to the most
simple and expected behaviour.